requirements.txt
asgiref==3.5.0
dj-rest-auth==2.2.3
Django==4.0.3
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5